requirements.txt
flask
scipy
scikit-learn
numpy
pandas 
pymysql
python-dotenv
mysql-connector-python
-e .